daily: reordering to dailies and Monthly calendar
devel level: dummy is now  separate directive
dailies are loaded first and then tasks are loaded.
there is a dirty bit for each of the affected dailies and a global dirty
bit
the tasks store a pointer to th daily thus type_order is available for
task which are never done before. (if a task was never done before its
listing will be empty as a result the type_order is never available to
even make the first request)
there is a new table called typeorder to keep track for types and their
order on the screen