preload data using promises;
promises store their return value so we cache the promisee their sever
responses are already cached. So if the promises are re-resolved it like
effectively getting server response so we can re-iterate over the
received and use it as we like.

So promises are cached in $scope.monthlyexps.promise[key] and key is
year-month.

However these promises need to be recreated once that months data is
changed. therefore on delete or edit the corresponding promise is
deleted and new data is fetched form the server

Accordingly to pre-load the data corresponding functions have been
changed.